refractor: add timer functionality to ConfirmDialog
