+++ packages/ui/src/components/dialog/ConfirmDialog.tsx
@@ -1,5 +1,5 @@
 import classnames from "classnames"
-import { FunctionComponent } from "react"
+import { FunctionComponent, useEffect, useMemo, useState } from "react"
 import { Classes } from "../../styles"
 import CloseIcon from "../icons/CloseIcon"
 import Dialog from "./Dialog"
@@ -11,7 +11,7 @@ const ConfirmDialog: FunctionComponent<{
     onClose: () => void
     onConfirm: () => void
     isConfirmDisabled?: boolean // if true, confirm button is disabled
-    confirmDisabledText?: string // if isConfirmDisabled is true, this text is shown on the confirm button
+    confirmDisabledUntil?: Date // if set, confirm button is disabled until this time (no need to use isConfirmDisabled in this case)
 }> = ({
     title,
     message,
@@ -19,8 +19,39 @@ const ConfirmDialog: FunctionComponent<{
     onClose,
     onConfirm,
     isConfirmDisabled = false,
-    confirmDisabledText,
+    confirmDisabledUntil,
 }) => {
+    const [secondsRemaining, setSecondsRemaining] = useState(0)
+
+    const isButtonDisabled = useMemo(() => {
+        return isConfirmDisabled || secondsRemaining > 0
+    }, [isConfirmDisabled, secondsRemaining])
+
+    const refreshTimerText = useMemo(() => {
+        return `${(Math.floor(secondsRemaining / 60) % 60)
+            .toString()
+            .padStart(2, "0")}:${(secondsRemaining % 60)
+            .toString()
+            .padStart(2, "0")}`
+    }, [secondsRemaining])
+
+    useEffect(() => {
+        // if confirm is disabled and confirmDisabledUntil is defined , start a timer to update the seconds remaining
+        if (!isButtonDisabled || !confirmDisabledUntil) return
+        const intervalId = setInterval(() => {
+            setSecondsRemaining(secondsRemaining - 1)
+        }, 1000)
+        // clear the interval when the component unmounts
+        return () => clearInterval(intervalId)
+    }, [secondsRemaining])
+
+    useEffect(() => {
+        if (!confirmDisabledUntil) return
+        setSecondsRemaining(
+            Math.ceil((confirmDisabledUntil.getTime() - Date.now()) / 1000)
+        )
+    }, [confirmDisabledUntil])
+
     return (
         <Dialog open={open} onClickOutside={onClose} className="px-6">
             <div>
@@ -55,19 +86,20 @@ const ConfirmDialog: FunctionComponent<{
                         </button>
                         <button
                             onClick={(e) => {
+                                if (isButtonDisabled) return
                                 e.stopPropagation()
                                 onConfirm()
                                 onClose()
                             }}
                             className={classnames(
                                 Classes.button,
-                                isConfirmDisabled &&
+                                isButtonDisabled &&
                                     "bg-gray-900 border-gray-900 opacity-50"
                             )}
-                            disabled={isConfirmDisabled}
+                            disabled={isButtonDisabled}
                         >
-                            {isConfirmDisabled && confirmDisabledText
-                                ? confirmDisabledText
+                            {isButtonDisabled && refreshTimerText
+                                ? refreshTimerText
                                 : "Confirm"}
                         </button>
                     </div>
