+++ packages/ui/src/routes/account/AllowancesPage.tsx
@@ -1,5 +1,5 @@
 import { BigNumber } from "@ethersproject/bignumber"
-import { useEffect, useMemo, useRef, useState } from "react"
+import { useEffect, useRef, useState } from "react"
 
 import {
     addNewApproveTransaction,
@@ -38,6 +38,8 @@ type allowancesToRevoke = {
     spenderAddress: string
 }[]
 
+const timeToDisableRefresh = 5 * 60 * 1000
+
 const AllowancesPage = () => {
     const history = useOnMountHistory()
 
@@ -57,36 +59,9 @@ const AllowancesPage = () => {
         }
     )
 
-    const timeToDisableRefresh = 5 * 60 * 1000
-
-    // check if passed 5 minutes since last refresh from the persisted data
-    const isRefreshDisabled: boolean =
-        persistedData.lastTriggered + timeToDisableRefresh > Date.now()
-
-    const [secondsRemaining, setSecondsRemaining] = useState(
-        Math.ceil(
-            (persistedData.lastTriggered + timeToDisableRefresh - Date.now()) /
-                1000
-        )
+    const refreshDisabledUntil = new Date(
+        persistedData.lastTriggered + timeToDisableRefresh
     )
-
-    const refreshTimerText = useMemo(() => {
-        return `${(Math.floor(secondsRemaining / 60) % 60)
-            .toString()
-            .padStart(2, "0")}:${(secondsRemaining % 60)
-            .toString()
-            .padStart(2, "0")}`
-    }, [secondsRemaining])
-
-    useEffect(() => {
-        // if refresh is disabled, start a timer to update the seconds remaining
-        if (!isRefreshDisabled) return
-        const intervalId = setInterval(() => {
-            setSecondsRemaining(secondsRemaining - 1)
-        }, 1000)
-        // clear the interval when the component unmounts
-        return () => clearInterval(intervalId)
-    }, [secondsRemaining])
 
     const [groupBy, setGroupBy] = useState<AllowancesFilters>(
         history.location.state?.groupBy || AllowancesFilters.SPENDER
@@ -137,13 +112,10 @@ const AllowancesPage = () => {
     }
 
     const refetchAllowances = async () => {
-        if (!isRefreshDisabled) {
-            setPersistedData({
-                lastTriggered: Date.now(),
-            })
-            await refreshTokenAllowances()
-            setSecondsRemaining(timeToDisableRefresh / 1000)
-        }
+        await refreshTokenAllowances()
+        setPersistedData({
+            lastTriggered: Date.now(),
+        })
     }
 
     const onFilterChange = (filter: AllowancesFilters) => {
@@ -219,8 +191,7 @@ const AllowancesPage = () => {
                         payload: { status: "success" },
                     })
                 }}
-                isConfirmDisabled={isRefreshDisabled}
-                confirmDisabledText={refreshTimerText}
+                confirmDisabledUntil={refreshDisabledUntil}
             />
             <WaitingDialog
                 status={status}
